requirements.txt
Flask==2.3.2
flask[async]
openai
requests
python-dotenv